feature: move signup and  login page to pages director and fix login link redirection
