+++ src/router/router.tsx
@@ -4,10 +4,10 @@ import route from "../constants/routes";
 import IdeaCard_1 from "../components/IdeaCard_1";
 import IdeaCard_3 from "../components/IdeaCard_3";
 import IdeaCard_2 from "../components/IdeaCard_2";
-import Workspace from "../components/Workspace";
+import Workspace from "../pages/Workspace";
 import AccessRequest from "../components/AccessRequest";
-import SignUp from "../components/SignUp";
 import SignIn from "../components/Login/Index";
+import SignUp from "../pages/SignUp";
 
 // add your routes here
 const router = createBrowserRouter([
